Move LogServer to riaps.utils.logging
