+++ bin/logserver.py
@@ -11,33 +11,10 @@ import sys
 import copy
 import pickle
 import rpyc
-from rpyc.utils.server import ThreadedServer
+from riaps.utils.logging import LogServer
 from riaps.consts.defs import *
 
 logFile = 'riaps-logserver.log'
-
-class LogService(rpyc.Service):
-    """RPyC service for collecting remote logs
-
-    This acts as a server which remote clients can connect to. The
-    Clients can then call exposed_handle with any LogRecords which
-    should be logged. Those LogRecords are then passed to the local
-    Python logger.
-    """
-    ALIASES = [const.logServiceName] # Registry name for the service
-    
-    STOPPING = False
-
-    def exposed_handle(self, data):
-        """Exposed method for remote clients to pass log records to
-
-        Args:
-            data (bytes): Pickled LogRecord
-        """
-        if not LogService.STOPPING:
-            record = pickle.loads(data)
-            log = logging.getLogger(record.name)
-            log.handle(record)
 
 def main():
     """Start a Log Server
@@ -62,15 +39,8 @@ def main():
     logger = logging.getLogger(__name__)
     logger.setLevel(logging.INFO)
 
-    # Create and configure rpyc service
-    server = ThreadedServer(LogService, auto_register=True, protocol_config=const.logServiceConfig)
-
-    # Create catch for kill signals
-    def exit_gracefully(signal, frame):
-        LogService.STOPPING = True
-        server.close()
-    signal.signal(signal.SIGINT, exit_gracefully)
-    signal.signal(signal.SIGTERM, exit_gracefully)
+    # Create instance of the LogServer
+    server = LogServer()
 
     logger.info("\nStarting RIAPS Log Server on port " + str(server.port) + " and saving files to " + logFile)
